Relocate _path_to_inlet() and _verify_bore_connectivity()
