+++ pygfunction/gfunction.py
@@ -7,7 +7,7 @@ from scipy.interpolate import interp1d as interp1d
 from scipy.constants import pi
 import time as tim
 
-from .boreholes import Borehole
+from .boreholes import Borehole, _path_to_inlet, _verify_bore_connectivity
 from .heat_transfer import thermal_response_factors
 
 
@@ -784,76 +784,3 @@ def _temporal_superposition(dh_ij, Q):
     for it in range(nt):
         Tb_0 += dh_ij[:,:,it].dot(Q[:,nt-it-1])
     return Tb_0
-
-
-def _path_to_inlet(bore_connectivity, bore_index):
-    """
-    Returns the path from a borehole to the bore field inlet.
-
-    This function raises an error if the supplied borehole connectivity is
-    invalid.
-
-    Parameters
-    ----------
-    bore_connectivity : list
-        Index of fluid inlet into each borehole. -1 corresponds to a borehole
-        connected to the bore field inlet.
-    bore_index : int
-        Index of borehole to evaluate path.
-
-    Returns
-    -------
-    path : list
-        List of boreholes leading to the bore field inlet, starting from
-        borehole bore_index
-
-    """
-    # Initialize path
-    path = [bore_index]
-    # Index of borehole feeding into borehole (bore_index)
-    index_in = bore_connectivity[bore_index]
-    # Stop when bore field inlet is reached (index_in == -1)
-    while not index_in == -1:
-        # Add index of upstream borehole to path
-        path.append(index_in)
-        # Get index of next upstream borehole
-        index_in = bore_connectivity[index_in]
-
-    return path
-
-
-def _verify_bore_connectivity(bore_connectivity, nBoreholes):
-    """
-    Verifies that borehole connectivity is valid.
-
-    This function raises an error if the supplied borehole connectivity is
-    invalid.
-
-    Parameters
-    ----------
-    bore_connectivity : list
-        Index of fluid inlet into each borehole. -1 corresponds to a borehole
-        connected to the bore field inlet.
-    nBoreholes : int
-        Number of boreholes in the bore field.
-
-    """
-    if not len(bore_connectivity) == nBoreholes:
-        raise ValueError(
-            'The length of the borehole connectivity list does not correspond '
-            'to the number of boreholes in the bore field.')
-    # Cycle through each borehole and verify that connections lead to -1
-    # (-1 is the bore field inlet)
-    for i in range(nBoreholes):
-        n = 0 # Initialize step counter
-        # Index of borehole feeding into borehole i
-        index_in = bore_connectivity[i]
-        # Stop when bore field inlet is reached (index_in == -1)
-        while not index_in == -1:
-            index_in = bore_connectivity[index_in]
-            n += 1 # Increment step counter
-            # Raise error if n exceeds the number of boreholes
-            if n > nBoreholes:
-                raise ValueError(
-                    'The borehole connectivity list is invalid.')
-    return
